Added create race test and moved exception handler to top level
